fix: Client-side prediction no longer glitchy
Turns out this issue was stemming from multiple problems:
1. TickHistoryBuffer pushed too many empty frames when setting a frame
in the future.
2. TickHistoryBuffer defaulted to a configurable value when no frame was
found - a cleaner solution is to just throw an exception. In addition,
the state cache is backfilled on start, so we already have enough cached
data to use from the start.
3. TickHistoryBuffer's Get didn't recognize the earliest known frame and
returned the default value instead
4. The server always simulated up to the current tick, even if it didn't
have input available. This meant that the server was purely speculating
but treated these states as authorative.