fix: Handle un/registering providers during simulation
The NetworkRewindManager is iterating providers during simulation, and
registering a provider changes that collection, which results in an
exception.
The solution was to collect these changes and apply them only on the
next frame, so the simulation can go on undisturbed. These changes are
taken into consideration *instantly*, just the backing collection
remains unchanged.